requirements.txt
asgiref==3.8.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
defusedxml==0.8.0rc2
Django==5.0.6
django-filter==24.2
django-jazzmin==3.0.0
django-modeltranslation==0.19.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Faker==25.9.1
idna==3.7
Markdown==3.6
oauthlib==3.2.2
pillow==10.3.0
psycopg==3.1.19
psycopg-binary==3.1.19
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
typing_extensions==4.12.1
urllib3==2.2.1
